Light Detection Obj + Script
I used the code from seeside project to create a camera (which is under the floor for now but I will probably disable it later as I continue working) that will print to the debug console if the light is on or if the light is off in the room you are sitting in. As of now the value that determines if the lights are on or off needs to be changed manually.